Cpu: add pextrb variant

+++ emulator/include/x64/checkedcpuimpl.h
@@ -472,7 +472,7 @@ namespace x64 {
         [[nodiscard]] static u128 roundpd(u128 src, u8 imm, SIMD_ROUNDING);
 
         [[nodiscard]] static u128 pmulld(u128 dst, u128 src);
-        [[nodiscard]] static u32 pextrb(u128 src, u8 order);
+        [[nodiscard]] static u8 pextrb(u128 src, u8 order);
         [[nodiscard]] static u32 pextrd(u128 src, u8 order);
         [[nodiscard]] static u64 pextrq(u128 src, u8 order);
         [[nodiscard]] static u128 pinsrb(u128 dst, u8 src, u8 order);
